Create files for tests of specific tabs
